+++ test/runtests.jl
@@ -3896,6 +3896,7 @@ end
 @testset "test_libxml2_testcases" begin include("test_libxml2_testcases.jl") end
 @testset "test_w3c" begin include("test_w3c.jl") end
 @testset "test_tokenizer" begin include("test_tokenizer.jl") end
+@testset "test_allocations" begin include("test_allocations.jl") end
 @testset "test_cursor" begin include("test_cursor.jl") end
 @testset "test_flatnode" begin include("test_flatnode.jl") end
 @testset "test_node_identity" begin include("test_node_identity.jl") end